Adjust the Docker Compose deployments to work with the latest binaries
The latest binaries no longer provide default port numbers for HTTP.
This means that we can simply let them bind on the same port as what's
used externally, which improves readability. The scheduler now uses two
gRPC servers: one for clients and one for workers. Change the port
number allocations because of that.

Remove the -ubuntu16-04 suffix from the scheduler container. There is
now a single scheduler that can be used for multiple kinds of workers. A
single scheduler process is thus sufficient.

As all configuration files now use Jsonnet, we can use includes to
prevent repetition between configuration files. Common fields (such as
the blobstore configuration) are now stored in common.libsonnet and are
included by the other configuration files.