Test the diff between docker-compose and k8s configs
There should be limited differences between the Docker Compose and
Kubernetes configurations for Buildbarn. Until there is an integration
test for the Kubernetes deployment, check that the differences between
the configurations is minimal. Every now and then, we should also
remember to manually deploy and make sure it actually works on
Kubernetes.