Reduce maximumMessageSizeBytes to a safer limit
In the past we had to set it pretty high, as we also had to account for
Tree messages that can become pretty big. As Tree messages are nowadays
processed in a streaming manner and have a separate size limit, it's
possible to reduce this.

The old value is quite unsafe, as it allows workers to execute actions
with a Command message up to 16 MiB in size. Those may contain a very
large number of output files, which may cause ActionResult to also
become very big. So big in fact that it becomes impossible to embed in
Remote Worker's Synchronize() request, or REv2's Execute() or
GetActionResult() responses.