Tidy up
It is better to avoid the ',' sed separator, as that is used in many
other listings of the command line (where it is typically removed).
For instance, to run a sed command taken from 'strace':

    [pid 732776] execve("/usr/bin/sed", ["sed", "-i",
    "s#\\(ghcr\\.io/buildbarn/bb-storage:\\)[0-9tz]*-[0-9a-f]*#\\120231111T202247Z-ece87ab#g",
    "README.md", "docker-compose/docker-compose.yml",
    "kubernetes/bb-namespace.yaml", "kubernetes/browser.yaml",
    "kubernetes/frontend.yaml", "kubernetes/kustomization.yaml",
    "kubernetes/scheduler.yaml", "kubernetes/storage.yaml",
    "kubernetes/worker-ubuntu22-04.yaml"], 0x55cbf0d29900 /* 56 vars */)
    = 0

We want to remove the commas from the arg array listing.
A pound symbol is easier to retain in the sed expression.